스토리 배경음, 보스 스테이지 배경음
