+++ src/engine/Sound.java
@@ -42,5 +42,7 @@ public enum Sound {
     BGM_LV4,
     BGM_LV5,
     BGM_LV6,
-    BGM_LV7
+    BGM_LV7,
+    BGM_STORY,
+    BGM_STORY2
 }
